# CUDA: fastdiv, launch bounds for mmvq + q8_1 quant