This PR uses the new `fastdiv` code from https://github.com/ggml-org/llama.cpp/pull/15715 for MMVQ + q8_1 quantization, also adds `__launch_bounds__` to the quantization.

<details>
<summary>Performance</summary>

| GPU           | Model           | FlashAttention     |     Microbatch size | Test     |     t/s master |     t/s bc7497a09 |     Speedup |
| :------------ | :-------------- | :----------------- | ------------------: | :------- | -------------: | ----------------: | ----------: |
| MI60 / MI50   | llama 1B Q4_0   | No                 |                   1 | pp512    |         324.58 |            331.09 |        1.02 |
| MI60 / MI50   | llama 1B Q4_0   | No                 |                   2 | pp512    |         630.34 |            617.04 |        0.98 |
| MI60 / MI50   | llama 1B Q4_0   | No                 |                   3 | pp512    |         550.57 |            536.87 |        0.98 |
| MI60 / MI50   | llama 1B Q4_0   | No                 |                   4 | pp512    |         676.31 |            658.36 |        0.97 |
| MI60 / MI50   | llama 1B Q4_0   | No                 |                   5 | pp512    |         829.86 |            842.40 |        1.02 |
| MI60 / MI50   | llama 1B Q4_0   | No                 |                   6 | pp512    |         728.63 |            714.23 |        0.98 |
| MI60 / MI50   | llama 1B Q4_0   | No                 |                   7 | pp512    |         791.68 |            804.37 |        1.02 |
| MI60 / MI50   | llama 1B Q4_0   | No                 |                   8 | pp512    |         899.37 |            902.24 |        1.00 |
| MI60 / MI50   | llama 8B Q4_0   | No                 |                   1 | pp512    |          86.12 |             86.92 |        1.01 |
| MI60 / MI50   | llama 8B Q4_0   | No                 |                   2 | pp512    |         159.63 |            165.93 |        1.04 |
| MI60 / MI50   | llama 8B Q4_0   | No                 |                   3 | pp512    |         138.24 |            138.07 |        1.00 |
| MI60 / MI50   | llama 8B Q4_0   | No                 |                   4 | pp512    |         163.47 |            162.29 |        0.99 |
| MI60 / MI50   | llama 8B Q4_0   | No                 |                   5 | pp512    |         192.49 |            198.98 |        1.03 |
| MI60 / MI50   | llama 8B Q4_0   | No                 |                   6 | pp512    |         177.00 |            172.99 |        0.98 |
| MI60 / MI50   | llama 8B Q4_0   | No                 |                   7 | pp512    |         190.66 |            199.64 |        1.05 |
| MI60 / MI50   | llama 8B Q4_0   | No                 |                   8 | pp512    |         216.54 |            216.92 |        1.00 |
| P40           | llama 1B Q4_0   | Yes                |                   1 | pp512    |         277.78 |            300.09 |        1.08 |
| P40           | llama 1B Q4_0   | Yes                |                   2 | pp512    |         564.38 |            567.23 |        1.01 |
| P40           | llama 1B Q4_0   | Yes                |                   3 | pp512    |         749.81 |            741.21 |        0.99 |
| P40           | llama 1B Q4_0   | Yes                |                   4 | pp512    |         754.37 |            707.49 |        0.94 |
| P40           | llama 1B Q4_0   | Yes                |                   5 | pp512    |         943.11 |            961.92 |        1.02 |
| P40           | llama 1B Q4_0   | Yes                |                   6 | pp512    |        1021.61 |           1023.67 |        1.00 |
| P40           | llama 1B Q4_0   | Yes                |                   7 | pp512    |        1031.67 |           1059.86 |        1.03 |
| P40           | llama 1B Q4_0   | Yes                |                   8 | pp512    |        1025.45 |           1063.12 |        1.04 |
| P40           | llama 8B Q4_0   | Yes                |                   1 | pp512    |          56.44 |             58.36 |        1.03 |
| P40           | llama 8B Q4_0   | Yes                |                   2 | pp512    |         111.77 |            113.15 |        1.01 |
| P40           | llama 8B Q4_0   | Yes                |                   3 | pp512    |         154.34 |            153.60 |        1.00 |
| P40           | llama 8B Q4_0   | Yes                |                   4 | pp512    |         159.86 |            156.38 |        0.98 |
| P40           | llama 8B Q4_0   | Yes                |                   5 | pp512    |         189.98 |            194.64 |        1.02 |
| P40           | llama 8B Q4_0   | Yes                |                   6 | pp512    |         203.27 |            203.96 |        1.00 |
| P40           | llama 8B Q4_0   | Yes                |                   7 | pp512    |         194.65 |            201.68 |        1.04 |
| P40           | llama 8B Q4_0   | Yes                |                   8 | pp512    |         200.50 |            211.52 |        1.05 |
| RTX 3090      | llama 1B Q4_0   | Yes                |                   1 | pp512    |         665.98 |            679.69 |        1.02 |
| RTX 3090      | llama 1B Q4_0   | Yes                |                   2 | pp512    |        1102.97 |           1127.48 |        1.02 |
| RTX 3090      | llama 1B Q4_0   | Yes                |                   3 | pp512    |        1542.74 |           1565.06 |        1.01 |
| RTX 3090      | llama 1B Q4_0   | Yes                |                   4 | pp512    |        1929.88 |           1902.27 |        0.99 |
| RTX 3090      | llama 1B Q4_0   | Yes                |                   5 | pp512    |        2092.90 |           2079.23 |        0.99 |
| RTX 3090      | llama 1B Q4_0   | Yes                |                   6 | pp512    |        2241.82 |           2306.79 |        1.03 |
| RTX 3090      | llama 1B Q4_0   | Yes                |                   7 | pp512    |        2449.68 |           2448.28 |        1.00 |
| RTX 3090      | llama 1B Q4_0   | Yes                |                   8 | pp512    |        2556.59 |           2655.01 |        1.04 |
| RTX 3090      | llama 8B Q4_0   | Yes                |                   1 | pp512    |         155.02 |            156.56 |        1.01 |
| RTX 3090      | llama 8B Q4_0   | Yes                |                   2 | pp512    |         277.36 |            279.89 |        1.01 |
| RTX 3090      | llama 8B Q4_0   | Yes                |                   3 | pp512    |         390.23 |            390.96 |        1.00 |
| RTX 3090      | llama 8B Q4_0   | Yes                |                   4 | pp512    |         468.95 |            470.48 |        1.00 |
| RTX 3090      | llama 8B Q4_0   | Yes                |                   5 | pp512    |         514.50 |            509.70 |        0.99 |
| RTX 3090      | llama 8B Q4_0   | Yes                |                   6 | pp512    |         529.37 |            541.57 |        1.02 |
| RTX 3090      | llama 8B Q4_0   | Yes                |                   7 | pp512    |         559.17 |            559.30 |        1.00 |
| RTX 3090      | llama 8B Q4_0   | Yes                |                   8 | pp512    |         570.67 |            583.77 |        1.02 |
| RTX 4090      | llama 1B Q4_0   | Yes                |                   1 | pp512    |         879.19 |            891.23 |        1.01 |
| RTX 4090      | llama 1B Q4_0   | Yes                |                   2 | pp512    |        1358.20 |           1362.47 |        1.00 |
| RTX 4090      | llama 1B Q4_0   | Yes                |                   3 | pp512    |        1972.75 |           1974.98 |        1.00 |
| RTX 4090      | llama 1B Q4_0   | Yes                |                   4 | pp512    |        2590.58 |           2610.67 |        1.01 |
| RTX 4090      | llama 1B Q4_0   | Yes                |                   5 | pp512    |        2892.37 |           2929.62 |        1.01 |
| RTX 4090      | llama 1B Q4_0   | Yes                |                   6 | pp512    |        3278.21 |           3323.54 |        1.01 |
| RTX 4090      | llama 1B Q4_0   | Yes                |                   7 | pp512    |        3612.01 |           3629.45 |        1.00 |
| RTX 4090      | llama 1B Q4_0   | Yes                |                   8 | pp512    |        3945.15 |           3960.70 |        1.00 |
| RTX 4090      | llama 8B Q4_0   | Yes                |                   1 | pp512    |         189.43 |            190.74 |        1.01 |
| RTX 4090      | llama 8B Q4_0   | Yes                |                   2 | pp512    |         334.86 |            336.98 |        1.01 |
| RTX 4090      | llama 8B Q4_0   | Yes                |                   3 | pp512    |         494.97 |            497.43 |        1.00 |
| RTX 4090      | llama 8B Q4_0   | Yes                |                   4 | pp512    |         655.17 |            658.22 |        1.00 |
| RTX 4090      | llama 8B Q4_0   | Yes                |                   5 | pp512    |         768.36 |            776.39 |        1.01 |
| RTX 4090      | llama 8B Q4_0   | Yes                |                   6 | pp512    |         899.40 |            899.70 |        1.00 |
| RTX 4090      | llama 8B Q4_0   | Yes                |                   7 | pp512    |        1003.74 |           1004.35 |        1.00 |
| RTX 4090      | llama 8B Q4_0   | Yes                |                   8 | pp512    |        1086.69 |           1091.22 |        1.00 |
| RX 6800       | llama 1B Q4_0   | No                 |                   1 | pp512    |         236.03 |            236.85 |        1.00 |
| RX 6800       | llama 1B Q4_0   | No                 |                   2 | pp512    |         451.63 |            453.87 |        1.00 |
| RX 6800       | llama 1B Q4_0   | No                 |                   3 | pp512    |         626.35 |            631.96 |        1.01 |
| RX 6800       | llama 1B Q4_0   | No                 |                   4 | pp512    |         758.88 |            758.65 |        1.00 |
| RX 6800       | llama 1B Q4_0   | No                 |                   5 | pp512    |         845.78 |            851.50 |        1.01 |
| RX 6800       | llama 1B Q4_0   | No                 |                   6 | pp512    |         919.45 |            912.94 |        0.99 |
| RX 6800       | llama 1B Q4_0   | No                 |                   7 | pp512    |         962.88 |            960.16 |        1.00 |
| RX 6800       | llama 1B Q4_0   | No                 |                   8 | pp512    |        1043.47 |           1010.94 |        0.97 |
| RX 6800       | llama 8B Q4_0   | No                 |                   1 | pp512    |          64.34 |             64.62 |        1.00 |
| RX 6800       | llama 8B Q4_0   | No                 |                   2 | pp512    |         121.92 |            121.55 |        1.00 |
| RX 6800       | llama 8B Q4_0   | No                 |                   3 | pp512    |         158.20 |            161.59 |        1.02 |
| RX 6800       | llama 8B Q4_0   | No                 |                   4 | pp512    |         177.57 |            175.08 |        0.99 |
| RX 6800       | llama 8B Q4_0   | No                 |                   5 | pp512    |         188.46 |            186.37 |        0.99 |
| RX 6800       | llama 8B Q4_0   | No                 |                   6 | pp512    |         196.97 |            194.63 |        0.99 |
| RX 6800       | llama 8B Q4_0   | No                 |                   7 | pp512    |         200.01 |            197.98 |        0.99 |
| RX 6800       | llama 8B Q4_0   | No                 |                   8 | pp512    |         198.43 |            200.38 |        1.01 |

</details>

`fastdiv` alone is consistently faster, the addition of `__launch_bounds__` is faster on average but always for batch size 1 which is the most important use case.