@JohannesGaessler 2025-09-05
I also enabled `fastmodulo` used in the MoE branch, thanks for pointing it out. I noticed that some extra care must be taken for values which are used in only one of the device code branches. Trying to call `init_fastdiv_values` for 0 results in an arithmetic exception in host code. I kept it as an error but added an assert.

<details>
<summary>Performance</summary>

| GPU           | Model                | FlashAttention     |     Microbatch size | Test     |     t/s master |     t/s 748c6a53c |     Speedup |
| :------------ | :------------------- | :----------------- | ------------------: | :------- | -------------: | ----------------: | ----------: |
| MI60 / MI50   | deepseek2 16B Q4_0   | No                 |                   1 | pp512    |         111.86 |            114.86 |        1.03 |
| MI60 / MI50   | deepseek2 16B Q4_0   | No                 |                   2 | pp512    |         154.03 |            153.79 |        1.00 |
| MI60 / MI50   | deepseek2 16B Q4_0   | No                 |                   3 | pp512    |         164.48 |            162.50 |        0.99 |
| MI60 / MI50   | deepseek2 16B Q4_0   | No                 |                   4 | pp512    |         200.42 |            199.34 |        0.99 |
| MI60 / MI50   | deepseek2 16B Q4_0   | No                 |                   5 | pp512    |         235.12 |            236.64 |        1.01 |
| MI60 / MI50   | deepseek2 16B Q4_0   | No                 |                   6 | pp512    |         222.68 |            221.66 |        1.00 |
| MI60 / MI50   | deepseek2 16B Q4_0   | No                 |                   7 | pp512    |         240.91 |            244.66 |        1.02 |
| MI60 / MI50   | deepseek2 16B Q4_0   | No                 |                   8 | pp512    |         274.86 |            275.63 |        1.00 |
| P40           | deepseek2 16B Q4_0   | No                 |                   1 | pp512    |          77.76 |             80.49 |        1.04 |
| P40           | deepseek2 16B Q4_0   | No                 |                   2 | pp512    |         117.58 |            118.81 |        1.01 |
| P40           | deepseek2 16B Q4_0   | No                 |                   3 | pp512    |         149.76 |            150.04 |        1.00 |
| P40           | deepseek2 16B Q4_0   | No                 |                   4 | pp512    |         167.27 |            166.13 |        0.99 |
| P40           | deepseek2 16B Q4_0   | No                 |                   5 | pp512    |         192.36 |            192.58 |        1.00 |
| P40           | deepseek2 16B Q4_0   | No                 |                   6 | pp512    |         208.26 |            208.44 |        1.00 |
| P40           | deepseek2 16B Q4_0   | No                 |                   7 | pp512    |         224.89 |            224.44 |        1.00 |
| P40           | deepseek2 16B Q4_0   | No                 |                   8 | pp512    |         232.20 |            233.23 |        1.00 |
| RTX 3090      | deepseek2 16B Q4_0   | Yes                |                   1 | pp512    |         193.69 |            198.53 |        1.02 |
| RTX 3090      | deepseek2 16B Q4_0   | Yes                |                   2 | pp512    |         190.24 |            189.92 |        1.00 |
| RTX 3090      | deepseek2 16B Q4_0   | Yes                |                   3 | pp512    |         245.66 |            245.62 |        1.00 |
| RTX 3090      | deepseek2 16B Q4_0   | Yes                |                   4 | pp512    |         301.49 |            297.95 |        0.99 |
| RTX 3090      | deepseek2 16B Q4_0   | Yes                |                   5 | pp512    |         351.92 |            351.26 |        1.00 |
| RTX 3090      | deepseek2 16B Q4_0   | Yes                |                   6 | pp512    |         399.18 |            402.66 |        1.01 |
| RTX 3090      | deepseek2 16B Q4_0   | Yes                |                   7 | pp512    |         449.73 |            447.73 |        1.00 |
| RTX 3090      | deepseek2 16B Q4_0   | Yes                |                   8 | pp512    |         488.72 |            487.75 |        1.00 |
| RTX 4090      | deepseek2 16B Q4_0   | Yes                |                   1 | pp512    |         263.50 |            267.06 |        1.01 |
| RTX 4090      | deepseek2 16B Q4_0   | Yes                |                   2 | pp512    |         283.48 |            285.52 |        1.01 |
| RTX 4090      | deepseek2 16B Q4_0   | Yes                |                   3 | pp512    |         369.63 |            371.67 |        1.01 |
| RTX 4090      | deepseek2 16B Q4_0   | Yes                |                   4 | pp512    |         457.25 |            458.50 |        1.00 |
| RTX 4090      | deepseek2 16B Q4_0   | Yes                |                   5 | pp512    |         529.74 |            532.14 |        1.00 |
| RTX 4090      | deepseek2 16B Q4_0   | Yes                |                   6 | pp512    |         606.53 |            610.54 |        1.01 |
| RTX 4090      | deepseek2 16B Q4_0   | Yes                |                   7 | pp512    |         676.57 |            678.63 |        1.00 |
| RTX 4090      | deepseek2 16B Q4_0   | Yes                |                   8 | pp512    |         746.90 |            748.84 |        1.00 |
| RX 6800       | deepseek2 16B Q4_0   | No                 |                   1 | pp512    |          81.20 |             81.93 |        1.01 |
| RX 6800       | deepseek2 16B Q4_0   | No                 |                   2 | pp512    |         102.97 |            103.31 |        1.00 |
| RX 6800       | deepseek2 16B Q4_0   | No                 |                   3 | pp512    |         136.40 |            136.92 |        1.00 |
| RX 6800       | deepseek2 16B Q4_0   | No                 |                   4 | pp512    |         163.75 |            164.06 |        1.00 |
| RX 6800       | deepseek2 16B Q4_0   | No                 |                   5 | pp512    |         181.78 |            181.82 |        1.00 |
| RX 6800       | deepseek2 16B Q4_0   | No                 |                   6 | pp512    |         202.20 |            203.18 |        1.00 |
| RX 6800       | deepseek2 16B Q4_0   | No                 |                   7 | pp512    |         219.51 |            219.02 |        1.00 |
| RX 6800       | deepseek2 16B Q4_0   | No                 |                   8 | pp512    |         235.59 |            234.33 |        0.99 |

</details>